+++ NEWS.md
@@ -2,6 +2,10 @@
 
 ## PSMatch 1.13.2
 
+- Add possibility to split protein groups in PSM data with `makeAdjacencyMatrix`
+(see [issue #30](https://github.com/rformassspectrometry/PSMatch/issues/30))
+- Removed `showDetails()` from `setMethod("show", "PSM")` as discussed in 
+[issue #30](https://github.com/rformassspectrometry/PSMatch/issues/30)
 - Add `USI` parameter to `plotSpectraPTM`.
 - Correct `filterPsmFdr` output message
 - Update Fragments vignette.